new format of base desc
git-svn-id: http://localhost:8080/svn/AlgoQuest/trunk@82 de6f410c-f7d5-3442-bb8a-973c24c52ac3

+++ AQUtil/aq/FileMapper.cpp
@@ -58,7 +58,7 @@ int FileMapper::read(void * buffer, size_t offset, size_t len)
 	// Check if offset is in the current window
 	if ((pView == NULL ) || (offset < windowOffset) || ((offset + len) > (windowOffset + cbView)))
 	{
-		unsigned long long off_tmp = offset - (offset % cbView);
+		size_t off_tmp = offset - (offset % cbView);
 		this->remap(off_tmp);
 	}
